Use Coroutine Channels for Webapp communication
Replaces (overkill) IPC Messenger/Handler
