+++ app/src/main/java/org/jellyfin/mobile/bridge/NativePlayer.kt
@@ -1,22 +1,13 @@
 package org.jellyfin.mobile.bridge
 
 import android.content.Intent
-import android.os.Messenger
 import android.webkit.JavascriptInterface
 import org.jellyfin.mobile.MainActivity
 import org.jellyfin.mobile.player.ExoPlayerFormats
 import org.jellyfin.mobile.player.PlayerActivity
 import org.jellyfin.mobile.utils.Constants
-import org.jellyfin.mobile.utils.LifecycleAwareHandler
 
 class NativePlayer(private val activity: MainActivity) {
-
-    private val playerMessageHandler = LifecycleAwareHandler(activity.lifecycle) { message ->
-        val function = message.obj as? String
-        if (function != null) activity.loadUrl("javascript:$function")
-        true
-    }
-    private val webappMessenger = Messenger(playerMessageHandler)
 
     @JavascriptInterface
     fun isEnabled() = activity.appPreferences.enableExoPlayer
@@ -28,7 +19,6 @@ class NativePlayer(private val activity: MainActivity) {
     fun loadPlayer(args: String) {
         val playerIntent = Intent(activity, PlayerActivity::class.java).apply {
             putExtra(Constants.EXTRA_MEDIA_SOURCE_ITEM, args)
-            putExtra(Constants.EXTRA_WEBAPP_MESSENGER, webappMessenger)
         }
         activity.startActivity(playerIntent)
     }
