+++ app/src/main/java/org/jellyfin/mobile/player/PlayerViewModel.kt
@@ -12,27 +12,42 @@ import com.google.android.exoplayer2.SimpleExoPlayer
 import com.google.android.exoplayer2.analytics.AnalyticsCollector
 import com.google.android.exoplayer2.source.MediaSource
 import com.google.android.exoplayer2.util.Clock
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 import org.jellyfin.mobile.BuildConfig
+import org.jellyfin.mobile.WEBAPP_FUNCTION_CHANNEL
 import org.jellyfin.mobile.player.source.JellyfinMediaSource
 import org.jellyfin.mobile.player.source.MediaSourceManager
 import org.jellyfin.mobile.utils.*
 import org.jellyfin.mobile.utils.Constants.DEFAULT_SEEK_TIME_MS
 import org.jellyfin.mobile.utils.Constants.SUPPORTED_VIDEO_PLAYER_PLAYBACK_ACTIONS
+import org.koin.core.KoinComponent
+import org.koin.core.inject
+import org.koin.core.qualifier.named
 
-class PlayerViewModel(application: Application) : AndroidViewModel(application), LifecycleObserver, Player.EventListener {
+class PlayerViewModel(application: Application) : AndroidViewModel(application), KoinComponent, LifecycleObserver, Player.EventListener {
     val mediaSourceManager = MediaSourceManager(this)
 
     private val _player = MutableLiveData<ExoPlayer?>()
     private val _playerState = MutableLiveData<Int>()
     private var shouldPlayOnStart = false
 
+    /**
+     * Returns the current ExoPlayer instance or null
+     */
     val playerOrNull: ExoPlayer? get() = _player.value
 
     // Public LiveData getters
     val player: LiveData<ExoPlayer?> get() = _player
     val playerState: LiveData<Int> get() = _playerState
 
-    // Media session
+    /**
+     * Allows to call functions within the webapp
+     */
+    private val webappFunctionChannel: Channel<String> by inject(named(WEBAPP_FUNCTION_CHANNEL))
+    private var lastReportedPosition = -1L
+
     private val mediaSession: MediaSession by lazy {
         MediaSession(getApplication<Application>().applicationContext, javaClass.simpleName.removePrefix(BuildConfig.APPLICATION_ID)).apply {
             @Suppress("DEPRECATION")
@@ -89,9 +104,32 @@ class PlayerViewModel(application: Application) : AndroidViewModel(application),
         mediaSession.setMetadata(mediaSource.toMediaMetadata())
     }
 
-    private fun updateMediaSessionState() {
+    // Webapp interaction
+
+    private fun callWebAppFunction(function: String) {
+        webappFunctionChannel.offer(function)
+    }
+
+    private fun notifyEvent(event: String, parameters: String = "") {
+        callWebAppFunction("window.ExoPlayer.notify$event($parameters)")
+    }
+
+    private fun setupTimeUpdates() {
+        viewModelScope.launch {
+            while (true) {
+                updatePlaybackPosition()
+                delay(Constants.PLAYER_TIME_UPDATE_RATE)
+            }
+        }
+    }
+
+    private fun updatePlaybackPosition() {
         val player = playerOrNull ?: return
-        mediaSession.setPlaybackState(player, SUPPORTED_VIDEO_PLAYER_PLAYBACK_ACTIONS)
+        val playbackPositionMillis = player.currentPosition
+        if (player.playbackState == Player.STATE_READY && playbackPositionMillis > 0 && playbackPositionMillis != lastReportedPosition) {
+            notifyEvent(Constants.EVENT_TIME_UPDATE, playbackPositionMillis.toString())
+            lastReportedPosition = playbackPositionMillis
+        }
     }
 
     fun seekToOffset(offsetMs: Long) {
@@ -106,13 +144,30 @@ class PlayerViewModel(application: Application) : AndroidViewModel(application),
 
     @SuppressLint("SwitchIntDef")
     override fun onPlayerStateChanged(playWhenReady: Boolean, playbackState: Int) {
-        updateMediaSessionState()
+        val player = playerOrNull ?: return
         when (playbackState) {
             Player.STATE_READY -> {
                 mediaSourceManager.selectInitialTracks()
                 mediaSession.isActive = true
+                setupTimeUpdates()
             }
         }
+
+        // Update media session
+        mediaSession.setPlaybackState(player, SUPPORTED_VIDEO_PLAYER_PLAYBACK_ACTIONS)
+
+        // Update webapp state and position
+        when (playbackState) {
+            Player.STATE_ENDED -> {
+                notifyEvent(Constants.EVENT_ENDED)
+            }
+            else -> {
+                notifyEvent(if (player.isPlaying) Constants.EVENT_PLAYING else Constants.EVENT_PAUSE)
+                updatePlaybackPosition()
+            }
+        }
+
+        // Notify activity of current state
         _playerState.value = playbackState
     }
